Added option for calculation of MCLP objective function weights applied to each constraint that requires coverage (based on protection limit) for each scenario.
Previously this weight had been assigned the node population.  The implication was that if that node was not covered, the "damage" equalled the entire node population (or was proportional to it).  This is incorrect when the response at that node, when delivered the unprotected dose, is less than 100%.  The new option calculates this weight as:

w = Population*(ResponseIfUnprotected - MinimumResponseForCoverage)

where MinimumResponseForCoverage is the "protection limit" in the MCLP formulation.

Conceptually this weight selection more accurately reflects the damages avoided when covering a constraint and hence should steer the MCLP in a better direction wrt health impacts.  It remains to be seen if it matters.

The new weight selection is the default behavior (equivalent to using the -w1 flag).  To select population weighting, use the -w0 flag.  The readme file will be updated.

QA checks verify that after these changes, which required modification to the Con data structure, the design remains unchanged when using -w0.