The tso2Impact analysis module is now working to the extent that the intermediate analysis servers and the aggregation server now communicate and construct impact files, and do so using computations entirely within the original tso2Impact code.  In other words, any impact that appears in an impact file generated by the TEVA framework will have been computed by the same code used in the "tso2Impact" command-line tool. In further other words, the "tso2Impact" tool and the "tso2ImpactAnalysisModule.so" link in common utility libraries, so there's no duplication of the "processScenario" code, and changes therein are shared by both the standalone tool and the analysis module.
The original tso2Impact standalone code generates identical impact files for 
epanet example 3 ("test1" in packages/sp/test).  It still needs to be 
tested for the bigger networks to make sure nothing has changed.

The impact files generated by the TEVA analysis module aren't yet 
organized correctly, but the fact that the servers are creating them for 
the first time warranted a checkin.