Changes do not affect current operation; the following functionality is added:
It is useful to be able to generate a TSO file that includes only network header information (e.g. this can be used in Matlab using the MatlabTSO functions).  Previously this was not allowed because if you didn't specify a *.tsi or *.tsg the code would quit. Now it continues to run a hydraulic simulation, write the TSO header, and then close, leaving behind a TSO that is empty except for the header.