Added new EPANET Toolkit method ENcreatepattern().
This accepts a pattern ID and adds that to the internal pattern database.  This pattern ID is then valid to assign to a node demand or source, or for any other purpose that patterns are useful.  This new functionality is required to be able to do much with temporal uncertainty and variability, which would be represented withint EPANET by a random pattern.  

A typical sequence:

ENcreatepattern('NewPatID');
ENgetpatternindex('NewPatID',&NewPatIndex);
ENsetpattern(NewPatIndex,NewPat,N);
ENsetnodevalue(NodeIndex,EN_PATTERN,PatIndex);

where NewPat is a float pointer to N new pattern values.  In this example the new pattern is assigned to the demand of node 'NodeIndex'.