Updated to TSO version 9.
Version 9 fixes a potential bug in the index file when the any TSO file part is larger than 2GB
Version 9 is also compatible between 32 and 64 bit platforms.

Since version 8 has the potential bug when being written, any attempts to now create a version 8 file will cause an error.
Version 8 and earlier version  files should be able to be converted to version 9 via the tsotest utility program on the platform/architecture that created it rather than having to rerun the scenario.