Throughout the teva-sp distribution we've gotten rid of all occurrences of a certain word that begins with 'a' and describes intentional contamination.  This word has been replaced by 'event'.
There is potential confusion between this use of 
'event' and the use of that word in "event detection" research.  In 
teva-sp/SPOT, 'event' means that thing that begins with 'a' and might 
involve multiple injections at the same time.

I haven't touched the sp/grasp subdirectory since Jean-Paul is actively
developing there.  The occurrences of the 'a' word therein will have to
go when he's done.