We are now going to use "make install prefix={release dir}" to create binary release directories, which can then be tar'ed or zip'd.
So set up the tests in this directory so that they will be installed
into the proper place in the release directory.  Add an install hook
that changes the path to the sp script in the tests.  In the
development tree the path is ../../../bin/sp, and in the release
tree the path is ../bin/sp.