These are changes made to reduce the memory requirements of various spot tools.  In order to run one of the larger networks on EPA's cluster (where each machine only has 2GB of memory and is further constrained by mingw's 2GB limit), a careful analysis was made of the memory utilization of some of the tools.
The findings were that in some cases, the impact data was being loaded twice (and the unused format data was being kept in memory).  In other cases, the arrival sequence data was loaded and kept when it would not be used.  Also, due to the data structre used and the c++ vector expansion mechanism, there were cases where the vector growth operation would require too much memory.

The first two issues were addressed by specifying ways to opitionally load the appropriate data, either via a flag to the method or a NULL pointer passed in instead of the vector.  The last issue was addressed (for now) by calling the reserve() method so the vector did not need to grow.  In these cases, it would probably be better to have theses vectors hold pointers to the vector or map rather than the object (i.e vector<map<int,double> *>).

While these modifications sped up all uses of this code, and also enabled larger problems to be solved, there were still memory limits that could be further reduced.  The impact data, when loaded, could take up a significant amount of memory (in the case of heuristic representation=1, it would take numEvents * nnodes * 8 bytes, and even the sparse representation (=0) would not be much of an improvement if the data wasn't very sparse (likely due to overhead in the c++ map class).

To address this issue, I first simply dumped the entire 2-dimensional array of impact data to a file and every time a request was made for a value, the file pointer was moved to the appropriate location and the single value was read and returned.  This method is currently heuristic-representation=2.  This worked and required essentially no memory to store the impact data (only the FILE *), but was fairly slow.  I then added the ImpactCache class as described below and in the ImpactCache.h file.

In the PMImpactInstance class, two new types of impact file representation (--heuristic-representation) were added:

 2: The 2-dimensional impact data is stored in a binary file on disk that is used every time a request comes in for an impact value - an offset is calculated the file pointer is positioned to that offset and a single double value is read.  this method is fairly inefficient, but it allows for the greatest number of problems to be solved.
 3: The data is stored on disk in 2 files - one is a row-major representation and one is a colum-major representation of the impact data.  Entire rows or columns are stored in memory, but only a few at a time.   This takes advantage of access patters seen while "profiling" the randomsample executable.

With respect to the implementation of PMImpactInstance::readImpact, it would probably be best to have a superclass define the various methods for building and accessing the impact data and have subclasses for each representation (currently 4).  This would have the benefit of cleaning up the read_imapct code substantially and make it easier to read.

Other changes to the code in this commit are:
  added soming timing print statements to various pieces of code.
  extended the format of the impact file to include a second value on the first line of the file that holds the number of events stored in the file.  All the code that reads the impact file can currently handle either.  It was added because the ImpactCache NEEDS the number of events before it can be loaded, and if the value isn't there, it must read through the file twice.