Misc fix in tso2Impact headers.  A recent change in the semantics of the init() method was not properly propigated to subclasses.
Instrumented grasp examples to write out information in
a format that is easier to parse.

Setup the regression.study.xml file, and the corresponding
sp_test executable.  This format is much simpler than
the sp-regression.study.xml on which it is based.  This study
also includes tests that verify that the correct solution
value is found.