Modified to allow specification of the JDK root directory in order to correctly specify the include directories where the JNI-related headers can be found.
This is (at least in my opinion) a better way to do this rather than having to copy the proper files to the tevalibs/Module/include directory.

The option for the configure command is --with-jdk-root=JDK_ROOT_PATH.

If this is used on a cygwin/mingw install, the path cannot have any spaces or the configure command will fail.

These configuration changes have been test on cygwin/mingw and linux at ANL.