We need to create executables that are staticly linked.  If we use libtool at all, libtool is used to build all executables. It's all or nothing, we can not control how libtool is used.
The "-all-static" flag to libtool tells it to build everything
(including executables) as static.  The only way I can find right
now to get this flag to libtool is to add it to the LDFLAGS for
every executable individually:

grasp_LDFLAGS=-all-static

This is a real pain.  There should be an autoconf/libtool macro
that lets you set these flags.  Can't set global LDFLAGS, because
it breaks configure (which tries to compile programs with it).