we use "make install DESTDIR=<staging directory> to create binary distribution
we had an install hook in the Makefile that deleted all the acro-pico 
binaries and libraries that got installed but which we don't 
want in the
distribution.  this is not maintainable- always have to update this
list with what you *don't* want.

Now instead, the install hook deletes all acro-pico files found in
the installed bin and lib directories except for what we *do* want.
More maintainable list: "PICO, ufl, ..."