New executable: "aggregateImpacts {config file name}"
sp will use this to setup heuristic and ufl solvers to use
aggregation

It reads the configuration file and impact files, aggregates
locations, and writes a new configuration file and impact files
with superlocations replacing the original locations.

Fixed locations in original network become fixed locations in the
new network.

Invalid locations in the original network are filtered out.

Location costs in the original network are not handled yet.

The new configuration file has a line for each superlocation,
listing for each goal/event the original location that had the
worst impact.

Changes to SPProblem to support aggregateImpacts.

Two fixes to SPProblem::compute_superlocations:  We were 
supposed to save the worst original location for each 
superlocation, but we were saving the best instead.  Also
there was a divide by zero problem when we had zero impacts
that resulted in not enough aggregation in some cases.