Added the now-necessary flag --disable-aggregation=all.  Some tests have been failing since the new witness aggregation for heuristics was added. Comparing against old (non-aggregated) values was an error.
Also removed attempts to run "picoamp," as ampl is not on all of the platforms
on which this test runs.

In manual trials, the only test that fails is a lagrangian run that covers 
all events with 6 sensors in a 5-sensor run.  This is a known issue with 
Lagrangian that will be addressed in the future.