The PMem structure, which is initialized with a call to loadHIAOptions, contains information about the water network, the scenario, and the health impacts.  To make it possible for users of PMem (like tso2ImpactModule) to be compiled either with or without health impact information, we need to reorganize PMem.
The PThreat structure is a subset of the PMem structure that contains
no health impact information, and it is initialized by new function
loadThreatOptions.

Next I'll rewrite the PMem structure and loadHIAOptions to use handle
the sensitive part of the initialization, and to call loadThreatOptions
for the non-sensitive part of the initialization.

I'm doing this in the trunk rather than the reorg branch, because if
work continues in the trunk without this in place it will be hard to
merge into the reorg branch.  I'll also put this in the reorg branch.