tso2Impact's impact detection was pulled out into a detection library in packages/detect
For now, tso2Impact uses a detection object from the library that
does exactly what processScenario was doing before - a location
detects the contaminant when the species 0 water quality rises
above a minimum quality.

The idea is that we can substitute other detection schemes if
we want to.

For now, for a water quality measure, processScenario uses the
species 0 value read in from the tso file.  But maybe in the
future we would want the detection module to provide a single
value quality measure based on multi species numbers or some
other kind of analysis.  In that case, processScenario would
query the detection object for the water quality at a given time
step.