aggregateImpacts reads in an impact file and creates a smaller impact file by aggregating locations.  Previously, it read the entire impact file into memory before computing the output impact file, which is not necessary because the input impact file is processed in scenario number order.
aggregateImpacts now has an option to read the input impact in
chunks.  The option specifies the maximum number of impacts to
read in at one time.  (If a scenario has more impacts than the
maximum, the entire scenario will be read in.)

aggregateImpacts defaults to reading in the entire impact file
at once. 

sp defaults to invoking aggregateImpacts with a request to read
in 10000 impacts at a time

the new sp option --conserve-memory=NUM_IMPACTS can be used to
change this.  If NUM_IMPACTS is zero, the entire file will be
read in.