New class EventImpacts: This class manages an input impact file for all the applications (setupIPData, evalsensor, aggregateImpacts...) that use the SPProblem class.  An SPProblem object has a list of SPGoalInfo objects, each of which has an EventImpacts object.  The EventImpacts object can read the impact file in in chunks of contiguous scenarios.  It replaces the impacts in memory with new impacts when they are needed.  It is optimized for the case where applications read an impact file in scenario number order.
aggregateImpacts is the only application that will read the
impact file in chunks if an option is set for it to do that.
It defaults to reading in the whole file at once.

The rest of the applications use EventImpacts in it's
"read everything at once" mode.  They would need some work
to take advantage of the chunk-reading mode.  (setupIPData
uses global information from the impact file after it
is done processing it.  It could be changed to save the
data it needs as it's going along, but I haven't done
that.)