add "aggregateImpacts" to choice in the --memcheck option.  So --memcheck=all will run valgrind on aggregateImpacts along with all the other executables, or you can just --memcheck=aggregateImpacts to run valgrind on only aggregateImpacts
If --memmon is chosen, memmon is now also run on aggregateImpacts

When memmon was run on the solvers, the memmon result was displayed
along with the solver output, but when memmon was run on setupIPData,
aggregateImpacts, or preprocessImpacts, the memmon output was not
grabbed and printed.  Now, if --memmon is on, all of the output
from these three executables will be displayed, including the memmon
output.