Update to impact files.
Revision of GeneralSP.mod IP model to allow for the case where 
(a) 'nfd' is an intrinsict measure computed by this IP and
(b) the penalty for selecting the 'dummy' location is zero.

In particular, additional constraints were added to 
force detection when sensors are placed.  These constraints
are redundant when the penalty is >= the worst impact value.
Thus, the IP is not a p-median problem when the dummy
location has zero impact values!