Changed demand-based population from using actual demands from epanet to using the patterns specified in the file.  If the demands (as calculated by epanet) were used, there were a couple of issues:  1. if the simulation duration wasn't exactly 1 less than the least common multiple of all the pattern lengths in the inp file, the demands would not match.  2. TEVA's usage of the demands was modified a while ago to average concentrations over the reporting interval and this code also averaged the demands as well.  This again led to discrepancies when comparing to stand-alone runs or runs with different lengths.
Since the demands were only used to estimate population in the demand-based population case for HIA and to compute ingestion patterns for HIA, it really made sense to use the patterns instead of the computed demands anyway.  This data is computed in the function ENTU_CreateDemandPatterns when the stand-alone teva-sim is used, which calls new methods that had to be added to the EPANET toolkit to access all the pattern data.  The Java version computes the values itself and does not use the new EPANET methods (since it already loads the inp file data).

The TSO file was alos modified to store this new information.  It is currently stored in the prolog.

The HIA code was modified to use this new data in the appropriate caluclations as well.