In processScenario, updateImpactStatisticsDueToLink should be analygous to updateImpactStatisticsDueToNode, in that it should called for any non-zero impact at the Link.  We were calling it only if the impact exceeded the detectionLimit value.
Extent of Contamination is the only objective at this time that
uses updateImpactStatisticsDueToLink.

I also changed the prototype for updateImpactStatisticsDueToLink to
match updateImpactStatisticsDueToNode.  It didn't include the
water quality measure, and EC doesn't need that but some future
objective might.

This fixes the bug reported by Rob - that in the zero sensor
case the impacts differed if the detectionLimit was different.